refactor(exports): split dwca.py into package
Upcoming additions (Humboldt eco: terms, multimedia and MoF
extensions, EML 2.2.0) roughly double the code. Split by
responsibility now so each module has a single clear purpose.

Public API unchanged — re-exported from ami.exports.dwca.__init__.
